chore: rename folder and update imports
