+++ src/modules/user/adapters/secondary/userApiAdapter.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from "tsyringe";
 import { TYPES } from "@/di/types";
-import { type RestApiPort } from "@/modules/restApi/ports/secondary/restApiPort";
+import { type RestApiPort } from "@/modules/rest-api/ports/secondary/restApiPort";
 import { type UserApiPort } from "@/modules/user/ports/secondary/userApiPort";
 import { type GetUserResponseDto } from "@/modules/user/application/dtos/response.dto";
 import UserUrls from "@/modules/user/application/constants/userUrls";
